move the sql module up so non-test code can use it
